src/postoffice/requirements.txt
dnspython3==1.12.0
Flask==0.10.1
PyYAML==3.11